requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.5
sqlparse==0.4.3
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
google-auth==2.15.0
PyJWT==2.4.0
requests-oauthlib==1.3.1